requirements.txt
tensorflow==2.4.0
Pillow
selenium
selenium-wire